+++ src/ffi/mod.rs
@@ -1,8 +1,10 @@
 pub mod change_notifications;
+pub mod contact;
 pub mod core;
 pub mod format_and_print;
 pub mod properties;
 pub mod store;
+pub mod user_defaults;
 pub mod vcard_serialization;
 
 pub mod status {